requirements.txt
gs-quant
python-dotenv
autopep8